Extract Action from its package
