+++ src/main/java/stardeath/animates/participants/entities/Player.java
@@ -1,7 +1,7 @@
 package stardeath.animates.participants.entities;
 
 import stardeath.animates.Animate;
-import stardeath.animates.actions.Action;
+import stardeath.animates.Action;
 import stardeath.animates.participants.Faction;
 import stardeath.animates.visitors.AnimateVisitor;
 import stardeath.animates.visitors.DefaultAnimateVisitor;
